requirements.txt
pyside6
pyyaml
